creating weapon simplified
