+++ src/main/java/com/solank/fxglgames/sg/collision/BulletNoiseCollisionHandler.java
@@ -3,8 +3,8 @@ package com.solank.fxglgames.sg.collision;
 import com.almasb.fxgl.entity.Entity;
 import com.almasb.fxgl.physics.CollisionHandler;
 import com.almasb.fxgl.physics.PhysicsComponent;
-import com.solank.fxglgames.sg.Type;
 import com.solank.fxglgames.sg.components.TallNoiseComponent;
+import com.solank.fxglgames.sg.model.Type;
 
 import static com.almasb.fxgl.dsl.FXGLForKtKt.inc;
 import static com.almasb.fxgl.dsl.FXGLForKtKt.play;
