+++ src/main/java/com/solank/fxglgames/sg/collision/PlayerNoiseCollisionHandler.java
@@ -3,10 +3,9 @@ package com.solank.fxglgames.sg.collision;
 import com.almasb.fxgl.dsl.FXGL;
 import com.almasb.fxgl.entity.Entity;
 import com.almasb.fxgl.physics.CollisionHandler;
-import com.solank.fxglgames.sg.Type;
+import com.solank.fxglgames.sg.model.Type;
 
 import static com.almasb.fxgl.dsl.FXGL.getd;
-import static com.almasb.fxgl.dsl.FXGL.random;
 import static com.almasb.fxgl.dsl.FXGLForKtKt.inc;
 import static com.almasb.fxgl.dsl.FXGLForKtKt.play;
 
@@ -27,7 +26,7 @@ public class PlayerNoiseCollisionHandler extends CollisionHandler {
         b.removeFromWorld();
 
         //int randomInt = (int) (Math.random() * 2) + 1;
-        int randomInt = FXGL.random(1,2);
+        int randomInt = FXGL.random(1, 2);
 
         play("hit/hit" + randomInt + ".wav");
     }
