+++ src/main/java/com/solank/fxglgames/sg/components/TallNoiseComponent.java
@@ -29,7 +29,7 @@ public class TallNoiseComponent extends Component {
         freezeTimer = FXGL.newLocalTimer();
 
         this.upDown =
-            new AnimationChannel(FXGL.image("noise/tallNoise-up-down" + randomInt + ".png"), Duration.seconds(1), 3);
+                new AnimationChannel(FXGL.image("noise/tallNoise-up-down" + randomInt + ".png"), Duration.seconds(1), 3);
 
         this.texture = new AnimatedTexture(upDown);
         texture.setScaleX(2);
