+++ src/main/java/com/solank/fxglgames/sg/components/weapons/BigGun.java
@@ -3,7 +3,6 @@ package com.solank.fxglgames.sg.components.weapons;
 import com.almasb.fxgl.dsl.FXGL;
 import com.almasb.fxgl.entity.Entity;
 import com.almasb.fxgl.texture.Texture;
-import com.solank.fxglgames.sg.components.AbstractWeapon;
 import javafx.geometry.Point2D;
 import javafx.scene.image.ImageView;
 
@@ -11,6 +10,7 @@ public class BigGun extends AbstractWeapon {
 
     private final double COOLDOWN_DURATION = 1.5;
     private final double SHOT_PAUSE_DURATION = 0.55;
+    private final double DAMAGE = 20;
 
     public BigGun(Entity entity) {
         super(entity);
@@ -43,5 +43,10 @@ public class BigGun extends AbstractWeapon {
     @Override
     public double getShotPauseDuration() {
         return SHOT_PAUSE_DURATION;
+    }
+
+    @Override
+    public double getDamage() {
+        return DAMAGE;
     }
 }
