+++ src/main/java/com/solank/fxglgames/sg/components/weapons/WeaponComponent.java
@@ -1,4 +1,4 @@
-package com.solank.fxglgames.sg.components;
+package com.solank.fxglgames.sg.components.weapons;
 
 import com.almasb.fxgl.entity.component.Component;
 import com.almasb.fxgl.entity.components.TransformComponent;
@@ -9,12 +9,12 @@ public class WeaponComponent extends Component {
     private WeaponStrategy weapon;
 
     public WeaponComponent(WeaponStrategy weapon) {
-    this.weapon = weapon;
+        this.weapon = weapon;
     }
 
     @Override
     public void onAdded() {
-       weapon.onAdded();
+        weapon.onAdded();
     }
 
     @Override
@@ -25,18 +25,24 @@ public class WeaponComponent extends Component {
     public Point2D getWeaponOuterPoint() {
         return weapon.getWeaponOuterPoint();
     }
+
     public double getCooldownDuration() {
         return weapon.getCooldownDuration();
     }
+
     public double getShotPauseDuration() {
         return weapon.getShotPauseDuration();
+    }
+
+    public WeaponStrategy getWeapon() {
+        return weapon;
     }
 
     public void setWeapon(WeaponStrategy weapon) {
         this.weapon = weapon;
     }
 
-    public WeaponStrategy getWeapon() {
-        return weapon;
+    public Double getDamage() {
+        return weapon.getDamage();
     }
 }
