+++ src/main/java/com/solank/fxglgames/sg/model/Type.java
@@ -1,4 +1,4 @@
-package com.solank.fxglgames.sg;
+package com.solank.fxglgames.sg.model;
 
 public enum Type {
     TALLNOISE, NOISE, YUKINE, BULLET, WALL
