+++ src/main/java/com/solank/fxglgames/sg/ui/SGMainMenu.java
@@ -22,6 +22,7 @@ public class SGMainMenu extends FXGLMenu {
     private static final Color SELECTED_COLOR = Color.WHITE;
     private static final Color UNSELECTED_COLOR = Color.GRAY;
     private final ObjectProperty<sgButton> selectedButton;
+
     public SGMainMenu() {
         super(MenuType.MAIN_MENU);
         Texture bg = FXGL.getAssetLoader().loadTexture("mainmenu/mainmenu.png");
@@ -41,19 +42,19 @@ public class SGMainMenu extends FXGLMenu {
 
         var textDescription = FXGL.getUIFactoryService().newText("", Color.LIGHTGRAY, 14.0);
         textDescription.textProperty().bind(
-            Bindings.createStringBinding(() -> selectedButton.get().description, selectedButton));
+                Bindings.createStringBinding(() -> selectedButton.get().description, selectedButton));
 
         var box = new VBox(15,
-            btnPlayGame,
-            btnOptions,
-            new sgButton("placeholder1", "", () -> {
-            }),
-            new sgButton("placeholder2", "", () -> {
-            }),
-            btnQuit,
-            new Text(),
-            new Lineseperator(),
-            textDescription);
+                btnPlayGame,
+                btnOptions,
+                new sgButton("placeholder1", "", () -> {
+                }),
+                new sgButton("placeholder2", "", () -> {
+                }),
+                btnQuit,
+                new Text(),
+                new Lineseperator(),
+                textDescription);
 
 
         box.setTranslateX(100);
@@ -86,9 +87,9 @@ public class SGMainMenu extends FXGLMenu {
 
         public Lineseperator() {
             var gradient = new LinearGradient(0, 0, 1, 0, true, CycleMethod.NO_CYCLE
-                , new javafx.scene.paint.Stop(0, Color.WHITE)
-                , new javafx.scene.paint.Stop(0.2, Color.LIGHTGRAY)
-                , new javafx.scene.paint.Stop(1, Color.TRANSPARENT)
+                    , new javafx.scene.paint.Stop(0, Color.WHITE)
+                    , new javafx.scene.paint.Stop(0.2, Color.LIGHTGRAY)
+                    , new javafx.scene.paint.Stop(1, Color.TRANSPARENT)
             );
 
             line.setFill(gradient);
@@ -126,7 +127,7 @@ public class SGMainMenu extends FXGLMenu {
             text = FXGL.getUIFactoryService().newText(name, Color.WHITE, 20.0);
             text.fillProperty().bind(Bindings.when(focusedProperty()).then(SELECTED_COLOR).otherwise(UNSELECTED_COLOR));
             text.strokeProperty()
-                .bind(Bindings.when(focusedProperty()).then(SELECTED_COLOR).otherwise(UNSELECTED_COLOR));
+                    .bind(Bindings.when(focusedProperty()).then(SELECTED_COLOR).otherwise(UNSELECTED_COLOR));
 
             text.setStrokeWidth(0.5);
